Problem: non-consolidated validator state (fix #999)
Solution:
- validator-related state structures consolidated in "ValidatorState" and "ValidatorStateHelper"
- most direct sub-component updates or queries are replaced by operations on ValidatorState/ValidatorStateHelper that hide its internals
- verify_node_join decoupled from NodeInfo by taking a trait that provides the required information (as internals may change)
- cargo.lock updated to the new format (to reduce merge conflicts)
- tests updated to the new structures -- punishment.rs test had assertion changes from asserting on internals to asserting on overall abci outputs (like end block response)

NOTE: as I took over to work on #930, I find it hard to change / prepare the fix, as validator state interactions were scattered all around by individual component updates.
This PR is just some intermediate refactoring (in order to reduce coupling and increase cohesion;
the behaviour should be the same, but it may have fixed some small edge cases;
there's still a big room for improvement in future PRs), so that the fix for #930 is easier to do and the solution is easier to review, as the main logic would be in one place.